Implements delete functionality and component refactor
This commit implements the functionality to delete a booking entry.
It also fixes bug #2 where the table is not automatically being
refreshed on a successful delete or create operation.

The commit also refactors the implementation of composition in the
application. A "Booking" component is now created which passes the
parent state down to its children, which is the "React way" of state
management.